requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
joblib==1.4.2
mlflow==2.14.1
kafka-python==2.0.2
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.2
